dfgdf

+++ app/main.py
@@ -604,24 +604,24 @@ def start_publish(bot : telebot.TeleBot, user):
             bot.send_message(user, m_texto("La Operación ha finalizado") )
 
         
-
-        if scrapper.temp_dict[user].get("mostrar_tiempo_debug"):
-            
-            scrapper.temp_dict[user]["res"] = "\n".join(scrapper.temp_dict[user]["tiempo_debug"])
-
-            with open(os.path.join(user_folder(user), "tiempo_publicacion_" + str(user) + ".txt"), "w", encoding="utf-8") as file:
-                file.write("Log de publicación\nID del usuario: {}\n\n{}".format(user, scrapper.temp_dict[user]["res"]))
+        if scrapper.temp_dict.get(user):
+            if scrapper.temp_dict[user].get("mostrar_tiempo_debug"):
                 
-            with open(os.path.join(user_folder(user), "tiempo_publicacion_" + str(user) + ".txt"), "r", encoding="utf-8") as file:
-                bot.send_document(user, telebot.types.InputFile(file, file_name="tiempo_publicacion_" + str(user) + ".txt"), "Ha ocurrido un error inesperado! ID usuario: {}".format(user))
+                scrapper.temp_dict[user]["res"] = "\n".join(scrapper.temp_dict[user]["tiempo_debug"])
 
-        
+                with open(os.path.join(user_folder(user), "tiempo_publicacion_" + str(user) + ".txt"), "w", encoding="utf-8") as file:
+                    file.write("Log de publicación\nID del usuario: {}\n\n{}".format(user, scrapper.temp_dict[user]["res"]))
+                    
+                with open(os.path.join(user_folder(user), "tiempo_publicacion_" + str(user) + ".txt"), "r", encoding="utf-8") as file:
+                    bot.send_document(user, telebot.types.InputFile(file, file_name="tiempo_publicacion_" + str(user) + ".txt"), "Ha ocurrido un error inesperado! ID usuario: {}".format(user))
 
-        
-            os.remove(os.path.join(user_folder(user), "tiempo_publicacion_" + str(user) + ".txt"))
+            
+
+            
+                os.remove(os.path.join(user_folder(user), "tiempo_publicacion_" + str(user) + ".txt"))
 
 
-        liberar_cola(scrapper, user, bot)
+            liberar_cola(scrapper, user, bot)
 
     return
